[clear] search emojiとget emojiをリファクタ
